+++ gamma-backend/src/main/java/com/gamma/backend/service/auth/AlumnoAuthStrategy.java
@@ -17,7 +17,7 @@ public class AlumnoAuthStrategy implements RoleAuthStrategy {
     }
 
     @Override
-    public boolean estaActivo(String dni, Long anioActivo) {
+    public boolean estaActivo(String dni, Integer anioActivo) {
         Alumno alumno = alumnoRepository.findByDni(dni);
         return alumno != null &&
             "ACTIVO".equalsIgnoreCase(alumno.getEstado()) &&
