+++ gamma-backend/src/main/java/com/gamma/backend/service/auth/RoleAuthStrategy.java
@@ -2,5 +2,5 @@ package com.gamma.backend.service.auth;
 
 public interface RoleAuthStrategy {
     String getRol(); // Ej: "ADMINISTRADOR"
-    boolean estaActivo(String dni, Long anioActivo);
+    boolean estaActivo(String dni, Integer anioActivo);
 }
